% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/cleanNames.R
\name{cleanNames}
\alias{cleanNames}
\alias{cleanNames,GolmDB-method}
\title{Clean Accession Names}
\usage{
cleanNames(object)

\S4method{cleanNames}{GolmDB}(object)
}
\arguments{
\item{object}{a \code{GolmDB} S4 class}
}
\value{
a \code{GolmDB} class
}
\description{
Clean Accession Names
}
