% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillAccessionObject.R
\name{fillAccessionObject}
\alias{fillAccessionObject}
\title{Create Accession Object}
\usage{
fillAccessionObject(x)
}
\arguments{
\item{x}{a list element from a split Golm database}
}
\value{
a \code{GolmAccession} object
}
\description{
Create Accession Object
}
