% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/searchDB-method.R
\name{searchDB}
\alias{searchDB}
\alias{searchDB,golmDB-method}
\title{Search Database}
\usage{
searchDB(object, pattern)

\S4method{searchDB}{golmDB}(object, pattern)
}
\arguments{
\item{object}{a `golmDB` S4 class}

\item{pattern}{a character string to search for}
}
\value{
a `golmDB` class containing on the accessions which matched the search pattern
}
\description{
Search Database
}
