% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,GolmAccession-method}
\alias{show,GolmAccession-method}
\alias{show,GolmDB-method}
\title{show-GolmAccession}
\usage{
\S4method{show}{GolmAccession}(object)

\S4method{show}{GolmDB}(object)
}
\arguments{
\item{a}{`GolmDB` object}
}
\description{
show-GolmAccession

show-GolmDB
}
