% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIdentifier.R
\name{findIdentifier}
\alias{findIdentifier}
\title{Find Compound Identifier}
\usage{
findIdentifier(object)
}
\arguments{
\item{object}{a \code{GolmAccession} object}
}
\value{
a \code{GolmAccession} object
}
\description{
Find Compound Identifier
}
