% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openGolmDB.R
\name{openGolmDB}
\alias{openGolmDB}
\title{Parse the Golm Database}
\usage{
openGolmDB(GolmInput)
}
\arguments{
\item{GolmInput}{a valid file path for a Golm database file in `MSP` format}
}
\value{
a `GolmDatabase` object
}
\description{
Parse the Golm Database
}
