% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/searchDB-method.R
\name{searchDB}
\alias{searchDB}
\alias{searchDB,GolmDB-method}
\title{Search Database}
\usage{
searchDB(object, pattern)

\S4method{searchDB}{GolmDB}(object, pattern)
}
\arguments{
\item{object}{a \code{GolmDB} S4 class}

\item{pattern}{a character string to search for}
}
\value{
a \code{GolmDB} class containing on the accessions which matched the search pattern
}
\description{
Search Database
}
